// src/store/app/mutation-types.ts
export const enum AppMutationTypes {
  SET_WEBRTCSESSION_ID = 'SET_WEBRTCSESSION_ID',
  SET_CHATSESSION_ID = 'SET_CHATSESSION_ID',
  SET_DEVICE = 'SET_DEVICE',
  REMOVE_ALL_DEVICES = 'REMOVE_ALL_DEVICES',
  PERWFORMANCE_NAVIGATION_TYPE = 'PERWFORMANCE_NAVIGATION_TYPE',
  SET_VIDEO_ERROR_STATE = 'SET_VIDEO_ERROR_STATE'
}